+++ components/31-molecules/checkboxes-dynamic/checkboxes-dynamic.config.js
@@ -1,0 +1,58 @@
+'use strict';
+
+const generateCheckboxes = (uid, i) =>{
+  let result = [];
+  for (let j = i; j--;) {
+    result.push({
+      label: `Checkbox option ${j}`,
+      id: `checkbox-${+j}-${uid}`,
+      value: j
+    });
+  }
+
+  return result;
+};
+
+module.exports = {
+  title: 'Checkbox dynamic',
+  context: {
+    id: 'input-checkbox',
+    label: 'Checkboxes',
+    field_description: 'Optional field description.<br> --- <br> Lorem ipsum dolor sit amet, consectetur adipisicing elit. Accusantium consectetur eveniet illo porro quis sint.',
+    field_message: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec laoreet, urna sit amet convallis rhoncus, felis ex.',
+    description: 'Description checkboxes.',
+    options: generateCheckboxes('default', 2)
+  },
+  variants: [
+    {
+      name: 'more-then-6',
+      context: {
+        options: generateCheckboxes('default', 20)
+      }
+    },
+    {
+      name: 'more-then-20',
+      context: {
+        options: generateCheckboxes('default', 21)
+      }
+    },
+    {
+      name: 'with-error',
+      context: {
+        id: 'input_text--error',
+        label: 'input-text',
+        modifier: 'error',
+        options: generateCheckboxes('with-error', 2)
+      }
+    },
+    {
+      name: 'with-success',
+      context: {
+        id: 'input-text--success',
+        label: 'input-text',
+        modifier: 'success',
+        options: generateCheckboxes('with-success', 2)
+      }
+    }
+  ]
+};